+++ src/common/components/CalculatedExchangeAmount/index.js
@@ -1,0 +1,1 @@
+export { CalculatedExchangeAmount } from './CalculatedExchangeAmount';