+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * *BREAKING:* Bump up okapi interfaces for `pieces` (3.0). Refs UINV-529.
 * Invoice line Subscription fields are not populated correctly on Invoice line pane. Refs UINV-534.
 * Include the currency in an invoice upon creation derived from a purchase order. Refs UINV-530
+* Display the exchanged calculated total amount when invoice in foreign currency. Refs UINV-522.
 
 ## [5.0.1](https://github.com/folio-org/ui-invoice/tree/v5.0.1) (2023-10-31)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v5.0.0...v5.0.1)